restructure: consolidate ui/ files
