+++ app/blog/page.test.tsx
@@ -28,7 +28,7 @@ vi.mock('@/ui/sections/blog-post-list', () => ({
 
 // Mock PageLayout to avoid rendering Header/Footer in tests
 // But preserve the main wrapper that PageLayout now provides
-vi.mock('@/ui/layout/main', () => ({
+vi.mock('@/ui/layout/page-layout', () => ({
   default: ({ children }: { children: React.ReactNode }) => (
     <main id="main" className="flex-auto flex flex-col">
       {children}
