+++ app/likes/page.tsx
@@ -6,7 +6,7 @@ import fetchItunesItems from '@/io/itunes/fetchItunesItems'
 import { env } from '@/io/env/env'
 import { DEFAULT_OG_IMAGE, SITE_NAME, SITE_URL, TWITTER_HANDLE, SITE_LOCALE } from '@/seo/constants'
 
-import PageLayout from '@/ui/layout/main'
+import PageLayout from '@/ui/layout/page-layout'
 import MediaSection from '@/ui/sections/likes-row'
 
 export const metadata: Metadata = {
