+++ app/not-found.tsx
@@ -1,4 +1,4 @@
-import PageLayout from '@/ui/layout/main'
+import PageLayout from '@/ui/layout/page-layout'
 import NotFoundContent from '@/ui/sections/not-found-summary'
 
 export default async function NotFound() {
