+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { type ReactElement } from 'react'
-import PageLayout from '@/ui/layout/main'
+import PageLayout from '@/ui/layout/page-layout'
 import Summary from '@/ui/sections/home-summary'
 import RecentWriting from '@/ui/sections/home-recent-writing'
 import JsonLdScript from '@/seo/json-ld/script'
