--- ui/layout/nav-types.d.ts
@@ -1,4 +1,0 @@
-export type NavItem = {
-  text: string
-  href: string
-}
